+++ grokmirror/__init__.py
@@ -104,19 +104,20 @@ def git_newer_than(minver):
     return version.parse(ver) >= version.parse(minver)
 
 
-def run_shell_command(cmdargs, stdin=None):
+def run_shell_command(cmdargs, stdin=None, decode=True):
     logger.debug('Running: %s', ' '.join(cmdargs))
 
     child = subprocess.Popen(cmdargs, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     output, error = child.communicate(input=stdin)
 
-    output = output.decode().strip()
-    error = error.decode().strip()
+    if decode:
+        output = output.decode().strip()
+        error = error.decode().strip()
 
     return child.returncode, output, error
 
 
-def run_git_command(fullpath, args, stdin=None):
+def run_git_command(fullpath, args, stdin=None, decode=True):
     if 'GITBIN' in os.environ:
         _git = os.environ['GITBIN']
     else:
@@ -131,7 +132,7 @@ def run_git_command(fullpath, args, stdin=None):
     else:
         cmdargs = [_git, '--no-pager'] + args
 
-    return run_shell_command(cmdargs, stdin)
+    return run_shell_command(cmdargs, stdin, decode=decode)
 
 
 def _lockname(fullpath):