+++ CHANGELOG.rst
@@ -1,3 +1,10 @@
+v2.0.2 (2020-10-06)
+- Provide pi-piper utility for piping new messages from public-inbox
+  repositories. It can be specified as post_update_hook:
+  post_update_hook = /usr/bin/pi-piper -c ~/pi-piper.conf
+- Add -r option to grok-manifest to ignore specific refs when calculating
+  repository fingerprint. This is mostly useful for mirroring from gerrit.
+
 v2.0.1 (2020-09-30)
 -------------------
 - fix potential corruption when migrating repositories with existing